Rename kraus.py to kraus_protocol.py
