+++ cirq-core/cirq/protocols/kraus_protocol.py
