+++ cirq-core/cirq/protocols/kraus_protocol_test.py
